[qwen3_next vl] support dense model
